Profil Sayfası güncellendi. User ve Admin Sayfası Giriş Sayfası Ayarlandı

+++ KutuphaneOtomasyon/Controllers/CartController.cs
@@ -34,23 +34,41 @@ namespace KutuphaneOtomasyon.Controllers
             var user = User.FindFirst(ClaimTypes.NameIdentifier);
             var userıd = Convert.ToInt32(user.Value);
             var book = _context.Books.FirstOrDefault(x => x.BookId == bookid);
+            if (userıd==null)
+            {
+                return NotFound();
+            }
             if (book == null)
             {
                 return NotFound();
             }
-            var CartCount = _context.CartItems.Where(x => x.UserId == userıd).Count();
-            if (CartCount >= 3)
+            var userCartItems = _context.CartItems.Where(x => x.UserId == userıd).ToList();
+            var existingCartItem = userCartItems.FirstOrDefault(x => x.BookId == bookid);//Sepete eklenen aynı kitapmı aynı kitapsa adetini artır sepette çift gösterme
+            var totalQuantityInCart = userCartItems.Sum(x => x.Quantity) + (existingCartItem != null ? 1 : 0);
+            ViewBag.totalQuantityInCart = totalQuantityInCart;
+            if (totalQuantityInCart >= 3)
             {
                 TempData["ErroCart"] = "Sepete en fazla 3 kitap ekleyebilirsin";
-                return RedirectToAction();
+                return RedirectToAction("Index", "Home");
             }
-            var cartitem = new CartItem
+            if (existingCartItem!=null)
+            {
+                 
+                existingCartItem.Quantity += 1;
+            }
+
+            else 
+            { 
+            
+                var cartitem = new CartItem
             {
                 BookId = bookid,
-                UserId = userıd
+                UserId = userıd,
+                 Quantity = 1
             };
 
             _context.CartItems.Add(cartitem);
+            }
             _context.SaveChanges();
 
             TempData["SuccesCart"] = "Kitap sepete eklendi";
@@ -64,27 +82,59 @@ namespace KutuphaneOtomasyon.Controllers
             var user = User.FindFirst(ClaimTypes.NameIdentifier);
             var userıd = Convert.ToInt32(user.Value);
 
-            var cartitem = _context.CartItems.FirstOrDefault(book => book.Id == bookid && book.UserId == userıd);
+            var cartitem = _context.CartItems.FirstOrDefault(book => book.BookId == bookid && book.UserId == userıd);
 
             if (cartitem != null)
             {
-                _context.CartItems.Remove(cartitem);
-                _context.SaveChanges();
+                if (cartitem.Quantity>1)
+                {
+                    cartitem.Quantity -= 1;
+
+
+                }
+                else 
+                {
+                    _context.CartItems.Remove(cartitem);
+                   
+                }
+                _context.SaveChanges();     
+
             }
             return RedirectToAction("Index","Home");
 
 
         }
 
-        public IActionResult CheckoutCartItems() 
+        public IActionResult CheckoutCartItems() //Sepetteki kitapları toplu ödünç alıyor
         {
             var user=User.FindFirst(ClaimTypes.NameIdentifier);
             var userId = Convert.ToInt32(user.Value);
+            var mybook = _context.BookLoans.Where(x => x.UserId == userId).ToList();//Kullanıcının üzerindeki kitap sayılarını çektil
+            var cartItems = _context.CartItems.Include(x => x.Book).Where(x => x.UserId == userId).ToList();//Sepetteki toplam sayı çekmek için listeyi alıypz
+            int totalCartBook = cartItems.Sum(x => x.Quantity);//sepetteki toplam kitap adet bilgisini alıyoruz
+            int totalbook = mybook.Count + totalCartBook;//seppeteki kitap adet sayısı ile kullanıcının üzerindeki  kitap adet sayısının toplamı alıyoruz
 
-            var cartItems=_context.CartItems.Include(x=>x.Book).Where(x=>x.UserId==userId).ToList();
+
+            
+
+            if (totalbook>=3)
+            {
+                TempData["ErroCart"] = "Sepete aldıklarınız ve üzerinizde  bulunan kitapların toplam adeti 3 geçemez";
+                return RedirectToAction("Index", "Cart");
+            }
+
+             
 
             foreach (var item in cartItems)
             {
+
+                if (item.Book.Quantity<=0)//Kitap stoğu yoksa alamasın veya kitap stoğu yoksa listede göstermicem kullanıcı kitabı göremeyecek ayarlanacak o
+                {
+                    TempData["ErroCart"] = $"{item.Book.BookName} kitap stoktada yok ödünç alamazsın";
+                    continue;
+
+                }
+
                 var loan = new BookLoan
                 {
                     UserId = userId,
